WIN32's scandir() emulation did not allocate enough memory for directory names (STR #263)
Fl::compose() did not handle special keys like backspace
properly (STR #293)

Fl_Choice did not clip its text when drawing using the plastic
scheme (STR #287)

Fl_Group incorrectly mapped the emacs CTRL keys to keyboard
navigation (STR #228)

Fl_File_Browser::load() didn't handle a NULL directory name (STR
#266)

64-bit library fixes (STR #261, )

The Fl_Valuator::format() function did not limit the size of the
number buffer (STR #268)


git-svn-id: file:///fltk/svn/fltk/branches/branch-1.1@3211 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_x.cxx
@@ -1,5 +1,5 @@
 //
-// "$Id: Fl_x.cxx,v 1.24.2.24.2.30 2003/06/15 04:27:35 easysw Exp $"
+// "$Id: Fl_x.cxx,v 1.24.2.24.2.31 2004/02/26 03:06:41 easysw Exp $"
 //
 // X specific code for the Fast Light Tool Kit (FLTK).
 //
@@ -1085,7 +1085,7 @@ void Fl_X::make_xid(Fl_Window* win, XVisualInfo *visual, Colormap colormap)
     }
 
     // Make it receptive to DnD:
-    int version = 4;
+    long version = 4;
     XChangeProperty(fl_display, xp->xid, fl_XdndAware,
 		    XA_ATOM, sizeof(int)*8, 0, (unsigned char*)&version, 1);
 
@@ -1268,5 +1268,5 @@ void Fl_Window::make_current() {
 #endif
 
 //
-// End of "$Id: Fl_x.cxx,v 1.24.2.24.2.30 2003/06/15 04:27:35 easysw Exp $".
+// End of "$Id: Fl_x.cxx,v 1.24.2.24.2.31 2004/02/26 03:06:41 easysw Exp $".
 //
